fix manifest of index.zip relocation
and support starting websocket in a separate thread (useful on mybinder).
