+++ src/ada/fem/results/common.py
@@ -14,7 +14,7 @@ from ada.config import logger
 from ada.core.guid import create_guid
 from ada.fem.formats.general import FEATypes
 from ada.fem.shapes.definitions import LineShapes, MassTypes, ShellShapes, SolidShapes
-from ada.visit.comms import send_to_viewer
+from ...visit.websocket_server import send_to_viewer, start_ws_server
 from ada.visit.gltf.graph import GraphNode, GraphStore
 from ada.visit.gltf.meshes import GroupReference, MergedMesh, MeshType
 from ada.visit.rendering.renderer_react import RendererReact
@@ -435,7 +435,7 @@ class FEAResult:
         from trimesh.path.entities import Line
 
         from ada.api.animations import Animation, AnimationStore
-        from ada.visit.comms import WsRenderMessage, start_ws_server
+        from ...visit.websocket_server import WsRenderMessage
         from ada.visit.utils import in_notebook
 
         from ...core.vector_transforms import rot_matrix
