+++ src/ada/visit/websocket_server.py
@@ -1,7 +1,12 @@
-import argparse
 import asyncio
 import io
+import os
+import pathlib
+import platform
 import re
+import sys
+import threading
+import time
 from collections import defaultdict
 from dataclasses import dataclass, field
 from multiprocessing import Queue
@@ -9,7 +14,12 @@ from multiprocessing import Queue
 import trimesh
 import websockets
 
+import ada
 from ada.config import logger
+
+_THIS_DIR = pathlib.Path(__file__).parent
+WEBSOCKET_EXE_PY = _THIS_DIR / "websocket_cli.py"
+PYGFX_RENDERER_EXE_PY = _THIS_DIR / "rendering" / "render_pygfx.py"
 
 
 def pretty_print_ports(client_origins):
@@ -119,8 +129,6 @@ class WebSocketServer:
         import base64
         import json
 
-        from ada.visit.comms import WsRenderMessage
-
         with io.BytesIO() as data:
             scene.export(file_obj=data, file_type="glb", buffer_postprocessor=animation_store)
 
@@ -174,23 +182,127 @@ def start_server(shared_queue: Queue = None, host="localhost", port=8765):
     _server.start()
 
 
-if __name__ == "__main__":
-    argparse = argparse.ArgumentParser()
-    argparse.add_argument("--port", type=int, default=8765)
-    argparse.add_argument("--origins", type=str, default="localhost")
-    argparse.add_argument("--host", type=str, default="localhost")
-    argparse.add_argument("--debug", action="store_true")
-    args = argparse.parse_args()
+def start_server_in_thread(host, port, client_origins: list[str], debug_mode):
+    origins = []
+    if client_origins is None:
+        origins.append("http://localhost:5173")  # development server
+        for i in range(8888, 8899):  # local jupyter servers
+            origins.append(f"http://localhost:{i}")
+        origins.append("null")  # local html
+    else:
+        origins = client_origins
 
-    origins_list = []
-    for origin in args.origins.split(";"):
-        if origin == "localhost":
-            origins_list.append("http://localhost:5173")  # development server
-            for i in range(8888, 8899):  # local jupyter servers
-                origins_list.append(f"http://localhost:{i}")
-            origins_list.append("null")  # local html
+    _server = WebSocketServer(host=host, port=port, client_origins=origins, debug_mode=debug_mode)
+
+    # Create a new thread that runs the server's start method
+    server_thread = threading.Thread(target=_server.start)
+
+    # Start the thread
+    server_thread.start()
+
+
+def start_external_ws_server(server_exe, server_args):
+    if server_exe is None:
+        server_exe = WEBSOCKET_EXE_PY
+
+    args = [sys.executable, str(server_exe)]
+    if server_args is not None:
+        args.extend(server_args)
+
+    args_str = " ".join(args)
+    logger.info("Starting server in separate process")
+    # Start the server in a separate process that opens a new shell window
+    if platform.system() == "Windows":
+        os.system(f"start cmd.exe /K {args_str}")
+    elif platform.system() == "Linux":
+        os.system(f"xterm -e {args_str}")
+    elif platform.system() == "Darwin":
+        os.system(f"open -a Terminal.app {args_str}")
+    else:
+        raise NotImplementedError("Unsupported platform: {}".format(platform.system()))
+
+
+@dataclass
+class WsRenderMessage:
+    data: str  # This will hold the Base64 encoded bytes
+    look_at: list[float, float, float] | None = None
+    camera_position: list[float, float, float] | None = None
+
+
+def send_to_viewer(
+    part: ada.Part | trimesh.Scene, host="localhost", port=8765, origins: list[str] = None, meta: dict = None
+):
+    if origins is None:
+        send_to_local_viewer(part, host=host, port=port)
+    else:
+        send_to_web_viewer(part, port=port, origins=origins, meta=meta)
+
+
+def start_ws_server(
+    host="localhost",
+    port=8765,
+    server_exe: pathlib.Path = None,
+    server_args: list[str] = None,
+    run_in_thread=False,
+    origins: list[str] = None,
+    debug_mode=False,
+    override_binder_check=False
+) -> WebSocketServer:
+    ws = WebSocketServer(host=host, port=port)
+
+    # Check if we are running in a binder environment
+    res = os.getenv('BINDER_SERVICE_HOST', None)
+    if res is not None and override_binder_check is False:
+        run_in_thread = True
+
+    if ws.check_server_running() is False:
+        if run_in_thread:
+            start_server_in_thread(host=host, port=port, client_origins=origins, debug_mode=debug_mode)
         else:
-            origins_list.append(origin)
+            start_external_ws_server(server_exe, server_args)
 
-    server = WebSocketServer(host=args.host, port=args.port, client_origins=origins_list, debug_mode=args.debug)
-    server.start()
+        while ws.check_server_running() is False:
+            time.sleep(0.1)
+
+    return ws
+
+
+def send_to_ws_server(
+    data: str | bytes, host="localhost", port=8765, server_exe: pathlib.Path = None, server_args: list[str] = None
+):
+    ws = start_ws_server(host=host, port=port, server_exe=server_exe, server_args=server_args)
+
+    ws.send(data)
+
+
+def send_to_local_viewer(part: ada.Part | trimesh.Scene, host="localhost", port=8765):
+    """Send a part to the viewer. This will start the viewer if it is not already running."""
+    with io.BytesIO() as data:
+        start = time.time()
+        if isinstance(part, trimesh.Scene):
+            part.export(data, file_type="glb")
+        else:
+            part.to_trimesh_scene().export(data, file_type="glb")
+        end = time.time()
+        logger.info(f"Exported to glb in {end - start:.2f} seconds")
+
+        send_to_ws_server(data.getvalue(), host=host, port=port, server_exe=PYGFX_RENDERER_EXE_PY)
+
+
+def send_to_web_viewer(part: ada.Part, port=8765, origins: list[str] = None, meta: dict = None):
+    """Send a part to the viewer. This will start the viewer if it is not already running."""
+
+    start = time.time()
+    data = io.BytesIO()
+    scene = part.to_trimesh_scene()
+    scene.metadata["extra_meta"] = meta
+    scene.export(data, file_type="glb")
+
+    server_args = ["--port", str(port)]
+    if origins is not None:
+        server_args.extend(["--origins", ";".join(origins)])
+    end = time.time()
+
+    logger.info(f"Exported to glb in {end - start:.2f} seconds")
+
+    send_to_ws_server(data.getvalue(), port=port, server_exe=WEBSOCKET_EXE_PY, server_args=server_args)
